# Add mimalloc to improve multithreaded performance