@daniel-story 2025-06-12
> The feeling is that it has always been possible to override the allocators, that various teams have done so, and that we should not add a .cpp file specifically for one case of this. Instead the project should add CMake variables (options) to inject the additional material (.h, .cpp, and target link library) to allow mimalloc or other allocators on an equal footing.

This is a little unclear, so could you please elaborate on the exact design you want?

For example, there needs to be a single source file somewhere in the repository that \#includes mimalloc's operator new/delete overrides. If they can't be in any file in this repository, where should the \#include live?