@daniel-story 2025-06-17
@dneto0 Not sure if you had a chance to read https://github.com/KhronosGroup/SPIRV-Tools/pull/6172#issuecomment-2968196871 yet, but without further information about exactly what kind of design you're looking for we're inclined to close this PR.

The primary motivation of this PR was to improve the poor performance of the SPIRV-Tools-shared.dll delivered in the Vulkan SDK. If a solution that achieves this primary goal (i.e., uses mimalloc or some other similarly well-performing allocator by default) is not going to be mergeable then we will not be able to continue work on this.