@dneto0 2025-06-18
My intent was to start with this branch, and implement the indirection through CMake variable settings.