@dneto0 2025-06-19
After further discussion internally, the risk is really from downstream folks accidentally relying on mimalloc behaviours, because Hyrum's law kicks in.   (https://www.hyrumslaw.com/ )   And then the risk that their functionality breaks when mimalloc is *not* used.   We still think it's good for mimalloc to be used as the default as coded in the PR, and that benefit likely outweighs the risk.  So I won't block this further.

I'll clone this branch and fix up the merge conflicts.  Thank you for your patience.